+++ include/soloud_biquadresonantfilter.hpp
@@ -41,11 +41,11 @@ class BiquadResonantFilterInstance : public FilterInstance
     explicit BiquadResonantFilterInstance(BiquadResonantFilter* aParent);
 
     void filterChannel(float*       aBuffer,
-                       unsigned int aSamples,
+                       size_t aSamples,
                        float        aSamplerate,
                        time_t       aTime,
-                       unsigned int aChannel,
-                       unsigned int aChannels) override;
+                       size_t aChannel,
+                       size_t aChannels) override;
 
   protected:
     enum FilterAttribute
