+++ include/soloud_duckfilter.hpp
@@ -35,9 +35,9 @@ class DuckFilterInstance : public FilterInstance
 {
   public:
     void filter(float*       aBuffer,
-                unsigned int aSamples,
-                unsigned int aBufferSize,
-                unsigned int aChannels,
+                size_t aSamples,
+                size_t aBufferSize,
+                size_t aChannels,
                 float        aSamplerate,
                 time_t       aTime) override;
