+++ include/soloud_freeverbfilter.hpp
@@ -51,9 +51,9 @@ class FreeverbFilterInstance : public FilterInstance
 
   public:
     void filter(float*       aBuffer,
-                unsigned int aSamples,
-                unsigned int aBufferSize,
-                unsigned int aChannels,
+                size_t aSamples,
+                size_t aBufferSize,
+                size_t aChannels,
                 float        aSamplerate,
                 time_t       aTime) override;
     ~FreeverbFilterInstance() override;
