+++ include/soloud_lofifilter.hpp
@@ -51,11 +51,11 @@ class LofiFilterInstance : public FilterInstance
 
   public:
     void filterChannel(float*       aBuffer,
-                       unsigned int aSamples,
+                       size_t aSamples,
                        float        aSamplerate,
                        time_t       aTime,
-                       unsigned int aChannel,
-                       unsigned int aChannels) override;
+                       size_t aChannel,
+                       size_t aChannels) override;
 
     explicit LofiFilterInstance(LofiFilter* aParent);
 };
